pyproject.toml
[project]
name = "stravatracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.94.0",
    "google-genai>=1.72.0",
    "groq>=1.1.2",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "requests>=2.33.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
